GreenHub/momoapi/greenhub_trx_flask/requirements.txt
flask
psycopg2-binary
python-dotenv
requests
ipykernel
mtnmomoapi
